requirements.txt
numpy==1.23.3
gym==0.25.2
pytorch==1.13.0+cu117
matplotlib==3.6.0